Fixed and improved three test methods
Fixed 'createSObjectTest8', 'createSObjectListTest7' and 'createSObjectTest4', which were trying to insert an user record and throwing an error which has no fix by Salesforce (issue link: https://success.salesforce.com/issues_view?id=a1p300000008ZX2AAM). Replaced those with Account and Contact relationships which should work on most orgs (hopefully!). The last method now queries the 'BusinessHours' object to assign to the Case, and splitted the account from the create method (because most of the time the account should be the same as the contact's or asset's).